fix(depType): depType should not only be code
